Bug 3849: Duplicate certificate sent when using https_port
The certificate file given with the "cert=" option it may contain a list 
of
certificates to be chained to the SSL client, for example intermediate
certificates.

The bug caused because in the certificates  chain we are storing also 
the
certificate of the port. This is works well for SSL-bump because squid
generates a certificate which uses the port certificate as CA 
certificate.
But in the case of https_port without bumping the port certificate is 
sent
twice, one as SSL server certificate and one as chained certificate.

This patch try to chain port certificate only when the sslbump is used.


This is a Measurement Factory project