Non https connectiona on SSL-bump enabled port may stuck
This is can be seen on skype when try to connect to server using an
SSL-bump enabled squid port. Squid try to bump the connection, waiting 
for ever
the ssl protocol header, and skype client waits for ever  an answer from 
the
server.

This patch sets the timeout to Config.Timeout.request (request_timeout)

This is a Measurement Factory project